Added logging and removed caching
Sprinkle more logging information which will show up in our Docker logs,
and make it easier to track the processing of jobs. Closes #29

Also remove the caching of boto resources so that we assume a role more
often, which addresses the Expired Token. Another way of handling this
would be to add some timelimit to the cache, or introduce a new object
that manages and catches errors better. But this seemed easiest to do
for now, and has been running in for over a day without a noticeable
problem. Closes #28

Also allow --model to be used when creating a job from the command line.

And add caching of whisper model loading so it doesn't take the hit each
time loading it from disk.