Adjust input and output location
Look for media files in a directory named for the job id in the root of
the bucket. Then put the Whisper results in the job id directory in a
subdirectory named "output".

This commit also removes the clean up of media files, assuming that the
SDR workflow step will do that after successfully downloading results.

Closes #33