Check media and log attributes
This commit uses ffprobe, which comes with ffmpeg (a Whisper
requirement), to check that the file being transcribed is an audio or
video file that ffmpeg understands. The format, duration and size of the
file is logged to make it a bit easier to analyze the speech-to-text
logs.

If an invalid file is supplied to the service it will raise an exception
indicating that the format was invalid. This ultimately gets caught and
logged.

Closes #48
Closes #31