Fix exception handling
When running speech_to_text.py from the command line it will execute
main() before the SpeechToTextException class has been defined! This is
not the case when running the unit tests which load the entire module
before calling main().

Also, the runs list needs to be defined prior to the media being
processed or else it will only include the Whisper run information for
the last file.

Closes #60

Refs #59