Add Github Action deploy
With this configuration a Github Action will run when a release tag
`release-YYYY-MM-DD` has been created during weekly dependency updates,
The action will build the Docker container and deploy it to the development
(qa in SDR) and staging AWS environments. This will allow the First
Responder for the week to test it using the speech-to-text integration tests.

When the tag has been tested and is ready for production a developer
will need to create a release in Github using the release tag. This will
cause a build and deploy to the production AWS environment. We may want
to think about automated ways for this to happen, but the "serverless"
nature of AWS Batch means there really isn't a server for capistrano
(what we use to do other infra deploys) to talk to.

The keys for the different environments need to be set as
[Github Action Secrets](https://docs.github.com/en/actions/security-for-github-actions/security-guides/using-secrets-in-github-actions).

- AWS_ACCESS_KEY_ID_DEVELOPMENT
- AWS_SECRET_ACCESS_KEY_DEVELOPMENT
- AWS_ECR_DOCKER_REPO_DEVELOPMENT
- AWS_ACCESS_KEY_ID_STAGING
- AWS_SECRET_ACCESS_KEY_STAGING
- AWS_ECR_DOCKER_REPO_STAGING
- AWS_ACCESS_KEY_ID_PRODUCTION
- AWS_SECRET_ACCESS_KEY_PRODUCTION
- AWS_ECR_DOCKER_REPO_PRODUCTION

Note: only dlss-ops has permission to see the keys for the speech-to-text
user in production. So for now this is commented out, until we actually
do need to run in production. Maybe this could be ticketed as follow on
work?

Closes #46