Add Batch
This PR includes a major refactoring of speech-to-text.py so that it can
be run via AWS Batch instead of as a daemon waiting on a TODO queue.

Batch takes care of provisioning EC2 instances with the desired system
resources (e.g. GPU) when there is work to do, and will scale those down
when there is no remaining work to do.

The repository includes a stand-alone Terraform configuration that can be used by
anyone to set up Batch but it will need to be folded into the DLSS Terraform repository.

Refs https://github.com/sul-dlss/terraform-aws/issues/1177