Install dependencies with uv
We've been installing Python dependencies with Pip, and not tracking
their versions. Since we've started using uv in some other
infrastructure team Python projects it makes sense to add here so that
speech-to-text can be tracked by infra-team's weekly dependency update
process.

Unlike pip, uv always installs system specific Python wheels. There
wasn't a wheel available for triton (an openai-whisper's dependency)
under Python3.8 so the installation of dependencies failed.

So, in addition to adding uv this PR also upgrades our base Docker image
from `nvidia/cuda:12.1.0-devel-ubuntu20.04` to
`nvidia/cuda:12.8.0-cudnn-devel-ubuntu22.04` which allows us to install
python3.10 when we `apt install python3`.

Since this significantly change Whisper's behavior I wanted to be able to
compare the VTT transcript output before and after the Docker image
change. I added the start of a benchmarking system that will allow us to
compare the output of a set of 22 SDR items, with a previous benchmark.
Ideally this benchmark would be human vetted, and actually represent a
ground truth for what we believe the transcript should be. But for the
time being it is simply a snapshot in time of what the transcript looked
like today. See the benchmark/README.md file for details.

Closes #80
Refs #65